fix: change enter number from 0-8 to 1-9
Refactor board into game and display classes. WIP
